Make mandelbrot benchmark more fair
When a log object was allocated on the stack in the mandelbrot
performance test, access to it would trigger a page fault that caused
the OS to map the entire image buffer to physical memory. The nop
benchmark did not have this advantage as there was no log object being
accessed. Instead, as the mandelbrot code proceeded it would trigger a
new page fault for each 4 KiB page it reached, for a total of 256 page
faults. This slowed down the nop benchmark to the point where it was
actually slower than the reckless benchmark.

See http://stackoverflow.com/questions/30152288/ for more discussion.