Exception-safe shared_input_queue
Use the std::optional proposal to allow the shared input queue to be
recreated when the log is closed and opened, without causing exception-
safety issues. Before we would just explicitly call the dtor and then
the ctor, but if the ctor would fail then we would leave the object in
an undefined state.