Prevent deadlock when thread-local input buffer becomes empty
If the circular thread-local input buffer becomes empty when the write
cursor is in the middle of the buffer, we end up with segments that are
both too small to fit the largest permitted input frame. The code would
assume that more space will clear up eventually and block indefinitely.