Windows support
Extensive redesign to support MS Windows. The thread-local buffers and the
boost lockless queue are replaced with a simpler and more efficient
"magic ring buffer".

The Tup files were converted to Lua to simplify the necessary
conditional compilation for different platforms. However, due to Tup
issue 288 (https://github.com/gittup/tup/issues/288) this never worked
properly. Since Lua seems to be the way forward for full control of Tup
we'll keep it this way anyway, but it would seem Tup is not a good fit
for a portable project and should be replaced with something else
(CMake?).

On Windows we use Visual Studio project files for building the
project. Gitignore files are no longer automatically generated by Tup
since we need to manually update them to ignore the files created by
Visual Studio.

Whereas the core library is ported, most of the benchmarks are not.