Increase duration of call_burst and periodic_calls benchmarks
Since input-buffer size was increased for Windows support (16x bigger)
it takes longer for it to fill up, so we increase the length of the
call_burst benchmark by approximately the same amount.

The periodic_calls benchmark is also increased because there is a
great deal of variance in the beginning of the test, that later
subsides. This was not the case before, but it's probably due to
running benchmarks on new hardware that doesn't let me disable
frequency scaling. It could also have something to do with the
redesign of the input queue.