BAD: Don't set CPU affinity in single-threaded benchmarks
Performance logging in the call_burst benchmark shows that setting the
CPU affinity for the invoking thread actually causes the both the
invoking thread and the background thread to run on the same CPU. Or, at
least, Linux does execute them simultaneously.

The call to input_buffer_empty_event_.signal() in
basic_log::wait_for_input() schedules the invoking thread which starts
to push a ton of log entries on the input queue. But signal() doesn't
return until the log has filled up again. In other words the behavior is
as if pthread_setaffinity_np affects the entire process, not just the
caller thread.

AND THIS is not strange since affinity is inherited. Go figure.

Need to revert this and be more careful in setting affinity...