Base output buffer capacity on input queue size
Benchmarks have shown that the small output buffer capacity in relation
to the recently increased input-queue size causes reduced performance,
probably because the output buffer fills up and has to be flushed
multiple times while processing a single batch of input entries. So we
attempt to avoid that by setting the output buffer size based on how
many entries can be in the input queue.