Don't cache read position in mpsc_ring_buffer
In an effort to avoid some cache-line ping pong, the ring buffer's read
position is cached in a separate variable that is not shared with the
producers. We only need to get its real value whenever we appear to have
run out of space, hence not reading a variable that other threads are
constantly writing to.

In theory this seems good but so far I haven't been able to show in
benchmarks that it makes much of a difference. On the contrary, the
mandelbrot benchmark actually runs slightly slower, probably because of
additional branching and logic that needs to run on each iteration.

So, for now the caching is removed.