Remove dependency on performance_log in production builds
performance_log is only used in development for troubleshooting
performance issues and shouldn't need to be exported for the Conan
build, but the feature toggle was in the performance_log module rather
than in reckless. To avoid this we create a proxy header for
performance_log within the reckless subdirectory that performs the
feature toggling.