In-source native build with Conan
The canonical build scripts for production builds are either the GNU
Makefile or Visual Studio projects, so use those in Conan instead of the
cmake adaptation. Having to install both CMake and Conan just to build a
package seems a bit overkill for such a simple project with no external
dependencies.

Also, since this conanfile is in the source tree, use the
exports_sources approach recommended for that scenario in the Conan
documentation.

Finally, we forbid compiler.libcxx=libstdc++ since that causes ABI
incompatibility.