Bug fix: correct frame-pointer wraparound when processing batch
basic_log::output_worker calculates frame positions by taking the
"front" address (current read pointer) from mpsc_ring_buffer and then
advancing that until it reaches the end of the batch. However, if the
batch size is large enough to go past the end of the ring buffer then no
wrap around is applied to the frame pointer.

This isn't strictly a buffer overrun because, due to the nature of the
ring buffer implementation, the same physical memory is mapped in a
second time at the end and we are in fact meant to go past the end in
situations where objects overlap the end. However, we can run into
trouble when an object was constructed in the first virtual memory-map
and then accessed through the second virtual-memory map. For plain old
data it works fine, but if the object has code that depends on the
object's this-pointer staying consistent (a reasonable expectation) then
formatting or destroying the object can lead to undefined behavior.

This surfaced as github issue #36 in which the libstdc++ implementation
of std::string had optimizations that would make use of the address of a
class member to determine which deallocation strategy must be used.

The fix is fairly simple: instead of using the frame pointer as loop
invariant we use the ring buffer's 64-bit position to determine when the
end is reached, and whenever we need to use the frame pointer we make
sure that it is consistently wrapped to always start in the first
virtual-memory block.